+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+**Added:**
+* Trails for flowmotion
+
 **Fixed:**
 * Command Menu cpu usage being a bit high.
 * Menu button tooltip not aligned with the adaptive size of the bottom bars.